return item complete

+++ kss-pwa-app/src/app/account/order-details/order-details.component.ts
@@ -28,12 +28,14 @@ export class OrderDetailsComponent implements OnInit {
   showBackButton : boolean = false;
   cancelOrder : boolean = false;
   getCancelReason : any;
-  cancelReasons : any;
+  reasons : any;
   cancelReasonId : any;
   additionalRemark = '';
   cancelSuccessful : boolean = false;
   cancelOrderFailureMsg : any;
   returnItem : boolean = false;
+  cancelItemsList : any = [];
+  quantity = 1;
   constructor(private appservice : AppServiceService,
               private route: ActivatedRoute,
               private router: Router,
@@ -112,12 +114,18 @@ export class OrderDetailsComponent implements OnInit {
     return this.appservice.isLoggedInUser();
   }
 
-    openCancelOrder(){
+  openCancelOrder(){
+    this.cancelItemsList = Object.assign([], this.order.items)
+    this.cancelOrder = true;
+    this.callGetReasonsApi();
+  }
+
+  callGetReasonsApi(){
     $('#cd-cart').addClass('overflow-h');
     this.unsubscribeGetCancelReason();
     // let url = this.appservice.apiUrl +  "/api/rest/v1/district-state"
-    if(this.account_service.cancelReasons){
-        this.cancelReasons = this.account_service.cancelReasons;
+    if(this.account_service.reasons){
+        this.reasons = this.cancelOrder ? this.account_service.reasons.cancel : this.account_service.reasons.return;
         this.cancelReasonId = 0;
     }
     else{
@@ -127,8 +135,9 @@ export class OrderDetailsComponent implements OnInit {
       this.getCancelReason = this.apiservice.request(url, 'get', {}, {}, false, 'observable').subscribe((response)=>{
         console.log("response from location api ==>", response);
         response.cancel.push({id : 0, value : '--Select--' });
-        this.account_service.cancelReasons = response.cancel;
-        this.cancelReasons = response.cancel;
+        response.return.push({id : 0, value : '--Select--' });
+        this.account_service.reasons = response;
+        this.reasons = this.cancelOrder ? this.account_service.reasons.cancel : this.account_service.reasons.return;
         this.cancelReasonId = 0;
         this.appservice.removeLoader();     
       },
@@ -137,7 +146,6 @@ export class OrderDetailsComponent implements OnInit {
         this.appservice.removeLoader();
       })
     }
-    this.cancelOrder = true;
     this.cancelOrderFailureMsg = '';
   }
 
@@ -151,6 +159,10 @@ export class OrderDetailsComponent implements OnInit {
     // this.cancelReasonId = parseInt(this.cancelReasonId);
   }
 
+  confirm(){
+    this.cancelOrder ? this.callCancelOrderApi() : this.callReturnOrderApi();
+  }
+
   callCancelOrderApi(){
     this.appservice.showLoader();
     let url = this.appservice.apiUrl +  "/api/rest/v1/user/order/" + this.order.order_info.order_id +"/cancel";
@@ -159,7 +171,6 @@ export class OrderDetailsComponent implements OnInit {
       reason : this.cancelReasonId,
       comments : this.additionalRemark
     };
-    body._token = $('meta[name="csrf-token"]').attr('content');
     console.log("cancel body ==>", body);
 
     this.apiservice.request(url, 'post', body , header ).then((response)=>{      
@@ -181,7 +192,44 @@ export class OrderDetailsComponent implements OnInit {
         this.cancelOrderFailureMsg = "Order Cancellation failed. Please check your Internet Connection";  
       }
       else{
-        this.cancelOrderFailureMsg = error.message;
+        this.cancelOrderFailureMsg = error.error.message;
+      }
+      this.appservice.removeLoader();
+    })  
+  }
+
+  callReturnOrderApi(){
+    this.appservice.showLoader();
+    let url = this.appservice.apiUrl +  "/api/rest/v1/user/sub-order/" + this.cancelItemsList[0].suborder_id +"/return";
+    let header = { Authorization : 'Bearer '+this.appservice.getCookie('token') };
+    let body : any = {
+      reason : this.cancelReasonId,
+      comments : this.additionalRemark,
+      varaint_id : this.cancelItemsList[0].variant_id,
+      quantity : this.cancelItemsList[0].quantity
+    };
+    console.log("cancel body ==>", body);
+
+    this.apiservice.request(url, 'post', body , header ).then((response)=>{      
+      $('#cd-cart').animate({
+          scrollTop: 0
+      });
+      this.getOrderDetails();
+      this.closeCancelOrder();
+      this.cancelSuccessful = true;
+      this.order.cancel_allowed = false;
+    })
+    .catch((error)=>{
+      $('#cd-cart').animate({
+          scrollTop: 0
+      });
+      console.log("error ===>", error);
+      this.closeCancelOrder();
+      if(error.status == 0){
+        this.cancelOrderFailureMsg = "Item return failed. Please check your Internet Connection";  
+      }
+      else{
+        this.cancelOrderFailureMsg = error.error.message;
       }
       this.appservice.removeLoader();
     })  
@@ -190,13 +238,31 @@ export class OrderDetailsComponent implements OnInit {
   closeCancelOrder(){
     $('#cd-cart').removeClass('overflow-h');
     this.cancelOrder = false;
+    this.returnItem = false;
     this.cancelReasonId = '';
     this.additionalRemark = '';
   }
 
-  openReturnItem(){
-    $('#cd-cart').addClass('overflow-h');
+  openReturnItem(item){
+    console.log("item to be returned ==>", item);
+    this.cancelItemsList = [item];
+    this.quantity = item.quantity;
     this.returnItem = true;
+    this.callGetReasonsApi();
+  }
+
+  openQuantityModal(){
+    $('#qty-modal-account').modal('show');
+    $("#cd-my-account").css("overflow-y", "hidden");
+    $('.modal-backdrop').appendTo('.angular-app');
+
+    $('#qty-modal-cart').on('hidden.bs.modal', function (e) {
+      $("#cd-cart").css("overflow-y", "auto");
+    })
+  }
+
+  updateQuantity(quantity){
+    this.cancelItemsList[0].quantity = quantity;
   }
   
 }
